Created Components Folder and moved some stuff
