+++ src/components/CharDetails.js
