+++ src/index.js
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import GlobalStyles from './GlobalStyles';
-import './index.css';
 
 ReactDOM.render(
   <React.StrictMode>
